Download is now enabled for multiple sites, undi
